Simplify SafeStringWrapper.__new__
This removes an attempt at sanitizing values before passing values on to
the super class constructor. This doesn't seem necessary as far as I can
tell (see the doctest for my best guess as to why this was done).
This was introduced in
https://github.com/galaxyproject/galaxy/commit/581977f112d86c9b01d951aa75ed98696ab064d8,
which fixed a security issue.

Removing this silences a lot of `object() takes no parameters`
deprecation warnings on python 2.