Refactor TagManager (back) to TagHandler.
The original TagHandler used trans and app and was what I think we convinced of as a "Manager" - something that provides business logic and ties app and trans to the model (and other lower layers of Galaxy). Since 7c86c5480bc03db4897e64da1abd4c41ea6056c3 the tag manager has not needed app or trans to do any of its work. It is a clean model manipulation utility. It also doesn't use or check security, it isn't a "Manager" in our sense. It is a model utility and so it belongs in galaxy.model I believe.

This also "fixes" the problem of galaxy.managers dependending on galaxy.model and then galaxy.model depending back on galaxy.managers. Hopefully it is clear the dependency chain should be one direction and that is galaxy.managers depends on galaxy.model and not vice versa.

In addition to allowing us to properly create a package out the models, this also fixes something needling me in the history import/export and model store branch - the tag manager is/was the only manager I needed to import/export histories and such - that was good evidence I think that it wasn't a manager and should have been in models all along.