Fix incorrect lddas being used as collection members
Adjust library folder api to return ldda_id on ld items
and use it for the obect passed to colelction creater
instead of the ld_id.

This bug would only materialize if you have Galaxy isntance that
has a library dataset with multiple versions. Which is the reason why
it eluded us for some time. That and missing tests.