correct counting of paired output
If a tool returns a paired collection the test for the expected
number of outputs seems to be wrong. It seems that the pair and
its contents are counted (the reported number is larger by 2 than
the expected number).

I added a test to show this.

Also added a test with automatically detected forward and reverse element
which surprisingly works.