Release ampq connection after initial connect
We need to call `connection.connect()` to avoid
`'kombu.transport.sqlalchemy.Message' is not mapped`` errrors.
This is fine for for sqlalchemy transports, but the connection
is still open if we don't release the connection.

This is a problem when acquiring the connection with rabbitmq
transports. So instead we release the connection and acquire
a blocking connection from the connection pool. This sets up
the mapping and avoids `'kombu.transport.sqlalchemy.Message' is not
mapped`.
Not sure if this will fix @natefoo's issue as I haven't exactly
been able to reproduce the same error, but acquiring a blocking
connection failed outright without releasing the connection, so this is
probably a step in the right direction.