Delegate config file watching to single handler
This should reduce IO especially on instances that use a lot of
handlers and/or on a remote file system.

The process that watches the tool conf files will now
be the active process that alphabetially compares highest.
If that process goes away or is superseded by a process that
compares higher now that process will take over.
If a change has been found a control task will be sent via
the kombu messaging system.

Overall this works nicely, but while stopping the current
config watcher there is a small break during which changes
to the config don't trigger a reload, this could be handled
more gracefully.