Remove handle_panel_update
We can just call AbstractToolBox.create_section directly as the new
install operation is atomic (thank you so much @guerler!).

We don't need to have other processes create a new section, that'll
happen when we reload the toolbox anyway. Avoids a threading issue in
some circumstances that leads to tools appearing outside of sections.
Originally I added the task message in
https://github.com/galaxyproject/galaxy/commit/51c586c264f2acd912060297804d61c2539f2483
because the install process was a two step process, where we'd first
create the tool panel section (maybe in thread1) and then place the tool
into the not-yet-existing panel section in another thread.

I tried fixing that up again in
https://github.com/galaxyproject/galaxy/commit/fb69afae28c434f35860ccaa4591e95fe3acfe91,
which I think was almost correct, but send_control_task should have been
run with `noop_self` to avoid confliciting threads updating the
sections. Anyway, this is much better now, I think.