Bring more directories under packages/ export.
Add quota to data/ (should be refactored to not use trans but is straight forward and the imports are fine for now), and add three new packages.

- auth
- authnz
- app

A lot more needs to be added to app but we need to de-tangle the circular imports between app and web stuff and webapp stuff. It is messier than I thought.