Make styleguide work for vue components.
All components are now loaded in the styleguide, however to document
them jsdoc comments need to be added and `.md` files for examples. The
Alert component is now documented as an example.

Unfortunately many of our components use axios to access the API
immediately on creation. I'm not sure how to document these. It seems
like a bad thing that they do this...