Fix package test
The cause of this is related to pulling tool_util out of tools:
https://github.com/galaxyproject/galaxy/pull/8279

The solution here is a temporary hack (add helper module, util.py, to
tests/, then make tests a package).

A better solution will be to separate test/unit/tools into tools/ and
tool_util/ - I intend to take care of it soon.