catch exceptions occuring in expression validators
rationale: in any case the user should not be bothered with python
exceptions, but they are still of interest for admins and tool
developers.

Hence I suggest in case of any exception to raise a value Error with
the validator's message (hence the user will see the message intended
by the validator) and to write an error to the log for admins and tool
developers.