Fix default involucro path
BaseAppConfiguration.root is now an absolute path, and so is
BaseAppConfiguration.data_dir, so the result of
`resolve_path(kwargs.get("tool_dependency_dir", "dependencies"), self.data_dir)`
is not "none" as it used to be, but `"<galaxy_data_dir>/none"` and so
the fallback will not be used.

This can be fixed by setting the default target dir to `dependencies`
and join it to self.data_dir, which defaults to `galaxy_root/database`
on classic installations.

Fixes the issues in https://github.com/galaxyproject/galaxy/pull/8195#issuecomment-513246346