data_meta filter fix
- The filter assumed a meta data value of `None` if the meta data is not
set. Therefore this
https://github.com/galaxyproject/galaxy/blob/8ac4dc65ab316763d01975bd1d1b10ca90ee1481/lib/galaxy/tools/parameters/dynamic_options.py\#L201
evaluated to True only for non-existent meta data keys. The check should
be be against the value given in the no_value attribute of the filter
specification. For the implementation I used the element_is_set function
which also just checked for the existence of the meta data. Now it does
what the name suggests.
- For lists and collections it depended on the multiple attribute of the
filter if all elements were considered. Since this was not documented I
changed it to always use all elements