[19.09] Fix default for docker_sudo in new jobs configs.
We implemented Docker support in a time when a lot of examples required sudo I think or maybe I was just very confused... but I've come to regret the default docker_sudo being on (True) by default. All example job configs for Docker for Galaxy now include setting that to false as boiler plate.

YAML based job configuration files give us the opportunity to change this for new configuration files I think. This attempts to change the default to False (which also matches Singularity) for new YAML based configuration files.