Replace resolve_path() with os.path.join()
The function was redundant: it checked whether its argument was an
absolute path, and if not, called os.path.join. This is exactly what
os.path.join already does (python 3.x and 2.7)

The only benefit of having a separate function is the ability to add
additinal processing to such calls - so when we need that, we can add
it. Meanwhile, it removes unnecessary function calls + makes the code
slightly less complex.

The function is not called from outside this module (not to be confused
with the resolve_path method of the mixin class), but it's duplicated in
webapps/reports/config.py and webapps/tool_shed/config.py (which are not
addressed in this commit).