Bug fix - add missing invocation state tracking, report link.
This is imperfect and rough but provides a bare minimum implementation of a bunch of important functionality that the absence of represents serious enough usability problems that I think it belongs in 19.09.

For admins, there is no way to track workflow invocations being scheduled currently and we routinely ask people to set the state on the workflow invocation column with SQL directly. This is a serious issue and this commit provides a new list in the admin interface of active invocations with state tracking (for steps scheduled and jobs executed) with a cancel button for each.

The same component that gets embedded in the grid for state tracking is also now added to the landing page after workflows are executed. Each invocation gets its own tracking widget - we don't name workflow invocations so the creation time is used to distinguish between invocations (we could probably iterate on this and do something with inputs?). Even if ephemeral, this provides users a way of tracking workflow invocation progress and notifying them if scheduling fails. Previously we had no such notifications. It also provides them a way to cancel workflow invocations (less of a bug then these other points, but still important).

Once the workflow has finished executing, an invocation report link will appear - again the lack of a link inside of the app to the invocation report added in 19.09. Previously there was no link in the UI for this report.