Do not overwrite existing attributes
This is a safeguard against introducing a bug. This edit prevents the
config loading code from overwriting the value of an existing attribute.

The config loading procedure automatically creates attributes of the
GalaxyAppConfiguration object for each property in the schema. If the
schema contains a property with the same name as an existing attribute
(e.g., `reloadable_options`, `schema`, etc.), that attribute will be
overwritten. Worse, if the call to `_set_config_base()` is moved before
the call to `_create_attributes_from_raw_config()`, the values of
`config_dir` and `data_dir` will be set to `None`, because those are
present in the schema. This fix prevents this from happening.