Resolve path: new_file_path (no test)
This has been tested manually. It is expected to work because the
modification is identical to other similar config options (e.g.
`file_path`). The only code affecting this attribute was the deleted
assignment stm that resolved its path w.r.t. `data_dir`. This is
redundant code as it is now done automaticlaly; it is tested on other
config options.

This cannot be (easily) tested like other options becasue the value of
`new_file_path` is overwritten for integration tests:
https://github.com/galaxyproject/galaxy/blob/release_19.09/test/base/driver_util.py#L159
(same for the dev branch).