Fix 13 integration tests for config values
This builds on top of #8793.

Most of the fixes involve removing the `config` prefix from the default
value specified in the schema. There are 2 exceptions:

---------------------
1.
The default values for `oidc_config_file` and `oidc_backends_config_file`
are `xml` files: that's what is specified as defaults in the schema, and
that's what the sample files are in `config/sample/`.  Aslo,
`authnz/managers.py` and `authnz/__init__.py` both refer to `xml` files.

However, the loaded values are `yml` files; that's done here:
https://github.com/galaxyproject/galaxy/blob/dev/lib/galaxy/config/__init__.py#L677
I've traced it to this commit:
https://github.com/galaxyproject/galaxy/commit/5c5197e84120551b551164f6b73eea2df260a157

I assume it should be `xml`, so I've changed those 2 lines.
---------------------
2. I've removed the `config/` previx from this assignment:
https://github.com/galaxyproject/galaxy/blob/dev/lib/galaxy/config/__init__.py#L686

I assume it doesn't belong there, unless the intention was to have it
inside `config/config/`? (because `_in_config_dir()` places that value
inside the `config_dir`).
---------------------

Also, for consistency across the file, I've slightly edited the wording
of the note about values to be resolved with respect to another option
(with or without using the `path_resolves_to` attribute).