Refactor database_connection config loading
1. Do not check kwargs for schema options: that has been done

2. Assign default only if not supplied by user (read previously from kwargs)

3. Use single quotes for consistency (no special need for double quotes)

4. Do not specify `None` as default in kwargs.get(): it is the default

5. Start removing periods after comments for consistency (Unless it's a
long sentence, or more than one sentence. I'm fine with vise-versa, but
then comments should be added where missing).

6. Schema: remove line from description of `database_connection` becasue
it is self-referential: there's no additional info in the config
documentation file other than this description. Linking to sqlalchemy
docs seems excessive.

7. `database_connection` is untestable with current setup as it is
overridden for test purposes in `base/driver_util.py`. I'm leaving this
as an empty test+comment as a reminder to address this in the future.