Don't use a thread local ToolShedRepositoryCache
this may have been a memory leak.Every thread would need to build the
cache again when first accessing it. Now that we eagerload the
relationships (#8791 only went into dev) we shouldn't need to keep the
cache thread local. This should be faster and more memory efficient, and
actually it looks like the thread-local cache didn't work as intended
before (writing tests now to verify it does work now).