Prevent potential bug in config loading; test
This ensures that a config option can be set to None regardless of its
datatype. It fixes the case when a config option's datatype has an
associated type converter (currently, int, float, bool): if an option's
value were set to None and the option's datatype were an int, the type
converter function would do `int(None)`, which would raise an error.

This also adds `str` to the type converters (in the unlikely case a
non-string value gets supplied for a string options).

The unit tests (a) expose this potential bug, and (b) ensure a falsy
value is not evaluated to `None`.