Ensure schema default is used as fallback
We don't need to load this option from kwargs: it's already loaded. But
we DO check if it was set in kwargs, and if not, we fall back to the old
IPython-named alternative. If that one is not set, we fall back to
schema default.

Also, this ensures that loading of values (for options defined in the
schema) from kwargs is centralized.