Do not set schema default for ftp_upload_dir_template
Because it is not used: the value loaded uses `os.path.sep` instead of
the hardcoded slash. So the default in the schema was misleading.

Also, do not test default value, because it is set dunamically.
(it worked previously because schema default used the same separator as
the platform on which the test was run).