Refactor _set_config_base; fix config_dir value
1. Split method body in two
2. Change assignment of config_dir when running from source
3. Ensure value of config_dir is an absolute path

I think splitting the method imporoves readability: its functionality
can be easily grouped into 2 parts: (1) finding and reading the main
config file (I've removed the comment because it was not quite in place
and, I think, not essential to understanding the code); and (2)
setting paths for core config properties based on part (1).

The change to config_dir addresses previous discussions:
https://github.com/galaxyproject/galaxy/pull/8894#issuecomment-547027750c
https://github.com/galaxyproject/galaxy/pull/921#discussion_r304351292

I could've changed the line in question:
https://github.com/galaxyproject/galaxy/pull/921#discussion_r304351292
to this:
`if self.config_file is None and `config_dir` is not in config_kwargs:`

However, in my opinion, such an edit would further obfuscate the logic
of assigning that value. The proposed solution, I think, although more
verbose, makes the code more straightforward: it is immediately clear
what values are assigned under what circumstances.

Finally, in the previous version `config_dir` can be a relative path if
`config_file` exists, but `config_dir` is set by the user (to a relative
path). That, I believe, was a bug.