Move Jobs updated time display logic to client.
Displays actual time now on mouseover in addition to elapsed time now.

I want this for two reasons - first to eliminate the the admin/jobs web controller, we really should be using the jobs API for this component. This brings us closer to having these two controllers in sync. Second, I want to have a date display component for the invocations grid after merge of #8699 - the time display in those grids is very rough currently.