Fix name and version for export of a previous workflow version
The API call to retrieve an old version of a workflow:

```
GET /api/workflows/<workflow_id>?version=<N>
```

was returning always the latest workflow name, while the name of a previous
workflow version could have an important meaning for the user.
Also, when retrieving version 0, the returned dictionary contained the
latest version number instead of 0.