[PROCEDURES] Add Sergey Golitsynskiy to committers group
Sergey Golitsynskiy has been a very active contributor to the Galaxy
code base and ecosystem (ansible-galaxy, sequence-utils, iuc) during the
last year. His contributions are of high quality and contain appropriate
tests and documentation and it is a pleasure to review Sergey's
contributions.

I propose that we add Sergey to the Galaxy committers group so that we
can benefit from his experience and skills.