Add base_image column to mulled-build-files
This will continue to work for single column files
or files that make use of the second and third column already
(build_number and name_override).
I don't think we should be using name_override,
but given the library nature and not knowing if anyone
except for biocontainers/multi-package-containers uses
this I chose to keep backwards compaitbility.

This does add support for meaningful column headers,
so we can write something like
```
\#base_image	targets	build_number
extended	package	10
```
and not include name_override.