Sort best package by upload time (lowest priority)
in addition to version and build number. This is necessary for getting
the extended-base annotation for qiime 1.9.1, which does not contain the
annotation in linux-64/qiime-1.9.1-np110py27_1.tar.bz2 while in the
newer linux-64/qiime-1.9.1-np112py27_1.tar.bz2 it is present. This
depends of course on `conda search` returning packages sorted by upload
time and is a little brittle. It does seem better than always picking
the oldest upload time as we do without this change.  Also looks in
meta.yaml for extended base annotation, since sometimes this isn't
present in info/about.json.