Use RenamedTemporaryFile for safer integrated_tool_panel_file writing
If not using the largely undocumented and debug-y `integrated_tool_panel_tracking_directory`
we should use RenamedTemporaryFile which creates a temp file in the
directory where `integrated_tool_panel_file` lives, and upon succesful
write rename the file to `integrated_tool_panel_file`.
We use this in other places in the code and it seems to work well
when multiple threads are potentially modifying the target file.