Prevent tool execution getting stuck in history when a chromInfo dataset is in paused state
If a chromInfo dataset is paused (or in other non-terminal states)
execution will not start until the dataset is ready, which may be never.
So instead we prefer datasets in OK state and ignore those that
will never be OK. In theory that may fix some odd issues where
jobs never start running.