Update the current_history method.
In this method, it was assumed that the first returned history
from the API endpoint would be the currently active history,
but that list is returned in create-time descending order. The
`/history/current_history_json` is an existing way to retrieve
the data the `current_history` method needs to return.