Make history import/export test more stable
We need to wait for the second HDA to be finished, otherwise
deleting the dataset wil cancel the job (if it is still running)
and then the dataset will be in discarded state (either because
that's what cancelling the job does, or if the export starts first
it's the import logic that sets state discarded if the file was not
exported).
This means sometimes the state will be OK and sometimes it would
be discarded. With this change the state will be OK.