Fix vcf_bgzip mimetype, allow cors, tbi from metadata element
The mimetype would be text/plain otherwise, which is not compatible with
byte-range requests. `allow_cors` doesn't quite seem to work in my
tests, but we'll need this for jbrowse.

Also skips an additional conversion by using the tabix index, which
again is helpful when loading datasets from a shared history.