Enable nested collections as inputs to workflow tests.
This ties together seemingly random threads to get this done and tested within Galaxy.

There is this function in galaxy.tool_util.cwl.util called galactic_job_json that parses a CWL-style "job" (the inputs for a tool or a workflow execution as a JSON object) and parses it into a representation for Galaxy - uploading inputs to Galaxy as needed. The interaction with the API was originally abstracted out as two functions so that the Galaxy test code (in the CWL branch) and the Planemo driver code could interact with the API in different ways. Planemo used bioblend (loosely) and the Galaxy code used the test framework's "populators" abstractions.

This PR creates a new a layer above galactice_job_json that has some extension points that the Planemo/bioblend layer can take advantage of but by default will just use the galaxy_interactor stuff already being used by tool tests. So this new layer doesn't depend on "populators" or bioblend and should allow important code to be shared between the Galaxy test framework, the CWL branch of Galaxy, and Planemo. Having to update the code in both places was a significant problem with the CWL branch.

Here is a commmit demonstrating how it simplifies the CWL branch (https://github.com/jmchilton/galaxy/commit/d9e774fbd302cb83427543bd46850cb5938f007a). I don't have a matching commit for Planemo but the code is located here - https://github.com/galaxyproject/planemo/blob/master/planemo/galaxy/activity.py#L181 - that will be simplified after this commit is utilized.

This commit establishes that layer, refactors a couple of tool tests to use it, and then extends it to implement nested collections.

Another positive upsot of this commit is there is now some in core test coverage of galactic_job_json - an important library function.