Add subworkflow_invocation_id to invocation step
This is useful if we want to drill down into invocation steps.
Without this we don't have a direct handle on the details of an
invocation step. This will enable nicer workflow test reports
and a drill-down UI (Assuming I can pick up enough Vue ...).

Considering that subworkflow testing without this is a pain
I may want to backport it to 20.05.