Fix and simplify built-in trim tool
In the previous version, specifying a negative end position would cause
a "rotating" behavior of the slice when the reverse index was larger
than the length of the string.
The fixed version also detects out of range column specifications
correctly and should be faster since it uses a set instead of a list
lookup and does int conversions up front instead of inside the loop.