Switch tool document cache to diskcache
This should be more portable across different python installations.
If the fallback dbm.dumb module was used (instead of dbm.gnu or
dbm/ndbm) startup could become very, very slow. diskcache uses sqlite
and plain files to store the cache. With this being more portable we
should be able to store this on cvmfs and have other Galaxy instances be
able to consume the cache.