Pass user to ObjectStorePopulator.
Not used yet in Galaxy core yet, but useful for applications where you want object store selection to be based on user in some way. This code was taken from https://github.com/galaxyproject/galaxy/pull/4840. Part of this is trying to reduce the number of files that branch touches to make review easier - but I'm confident this extension point is good regardless. Also it makes it clear we need to keep the user object in the picture when assigning the object store ID in the future.