Fix serialization of Authnz config
If you define `__eq__` (`__ne__` isn't necessary on python 3)
you also need to define `__hash__` if you want to use the JSONType
column, otherwise the parent object (a CloudAuthz instance) can't be
stored in the weakref dictionary where mutations to are tracked.

But I think in this case it's better to use the more explicit `.equals()`
method.