Respect ``use_shared_home="false"`` when running tools in container
When a tool doesn't specify a `profile` (or sets a `profile` < 18.01) but use
``use_shared_home="false"`` to ask to set a HOME within the job's directory
(normally because the underlying tool writes to the home directory), this
works as expected when resolving dependencies with conda.
But when running the same tool in a Docker container (e.g. via
`planemo test --biocontainers`), the job script didn't export a temporary
HOME into the container (which instead setting `profile="18.01"` correctly
does).

Also add several integration tests for `use_shared_home`, of which
`DockerizedJobsIntegrationTestCase.test_container_job_environment_explicit_isolated_home`
would fail without this fix.