Allow execution of workflows for certain tool updates
This works around workflows not being executable when versions change.
We usually don't break any parameters or connections, so when we bump a
version of a tool shipped with Galaxy we can add the tool to
`WORKFLOW_SAFE_TOOL_VERSION_UPDATES`. That will let users run the
workflow, while still maintaining a warning that the tool version
doesn't match.