Restore welcome html link for non-uwsgi setups
by restoring the symlink from lib/galaxy/webapps/base/static to static.

This is the default location for `/static/welcome.html` as set in https://github.com/galaxyproject/galaxy/blob/release_20.09/lib/galaxy/webapps/base/webapp.py#L1032).
This fixes the `Not Found` in the center panel for Selenium tests and
planemo.
I am not sure if this is the right fix w.r.t packaging.

Alternatively we could sett a default of `static/welcome.html` for `static_welcome_html`
(and putting this in the schema, I have the impression that some users think this is what
`welcome_url` does).