Fix updating workflow node input terminal
This fixes a silent error when hitting
`this.outputTerminals[name].$el.remove();`

This seems to not be necessary, but it is super weird that there is no
error logged. I just found it by stepping through this line by line in
the debugger.

This fixes https://github.com/galaxyproject/galaxy/issues/10622